requirements.txt
pymongo
python-dotenv
easyocr
pymupdf